Re-use project root computation in Rascal.
