+++ rascal-lsp/src/main/rascal/lsp/lang/rascal/lsp/IDECheckerWrapper.rsc
@@ -206,39 +206,9 @@ loc targetToProject(loc l) {
     return l;
 }
 
-@memo
 @synopsis{Infers the root of the project that `member` is in.}
-loc inferProjectRoot(loc member) {
-    parentRoot = member;
-    root = parentRoot;
-
-    do {
-        root = parentRoot;
-        parentRoot = inferDeepestProjectRoot(root.parent);
-    } while (root.parent? && parentRoot != root.parent);
-    return root;
-}
-
-@synopsis{Infers the longest project root-like path that `member` is in.}
-@pitfalls{Might return a sub-directory of `target/`.}
-loc inferDeepestProjectRoot(loc member) {
-    current = targetToProject(member);
-    if (!isDirectory(current)) {
-        current = current.parent;
-    }
-
-    while (exists(current), isDirectory(current)) {
-        if (exists(current + "META-INF" + "RASCAL.MF")) {
-            return current;
-        }
-        if (!current.parent?) {
-            return isDirectory(member) ? member : member.parent;
-        }
-        current = current.parent;
-    }
-
-    return current;
-}
+@javaClass{org.rascalmpl.vscode.lsp.rascal.model.ProjectRoots}
+java loc inferProjectRoot(loc member);
 
 map[loc, set[Message]] filterAndFix(list[ModuleMessages] messages, set[loc] workspaceFolders) {
     set[Message] empty = {};
