+++ rascal-lsp/src/test/java/org/rascalmpl/vscode/lsp/rascal/model/PathConfigsTest.java
@@ -43,10 +43,11 @@ public class PathConfigsTest {
     private static final IRascalValueFactory VF = IRascalValueFactory.getInstance();
     private static final URIResolverRegistry reg = URIResolverRegistry.getInstance();
 
+    private final ProjectRoots roots = new ProjectRoots();
 
-    private static void checkRoot(ISourceLocation project, String modulePath) throws URISyntaxException {
+    private void checkRoot(ISourceLocation project, String modulePath) throws URISyntaxException {
         var m = VF.sourceLocation(project.getScheme(), project.getAuthority(), project.getPath() + "/" + modulePath);
-        var root = PathConfigs.inferProjectRoot(m);
+        var root = roots.inferProjectRoot(m);
         assertEquals(project, root);
     }
